Refactor code that enables deterministic operation of cuDNN
