+++ tensorflow/core/kernels/BUILD
@@ -531,7 +531,6 @@ tf_cuda_library(
         "//tensorflow/core:lib",
         "//tensorflow/core:stream_executor",
         "//tensorflow/core/util/proto:proto_utils",
-        "//tensorflow/stream_executor/cuda:cuda_helpers",
         "//tensorflow/stream_executor/gpu:asm_compiler",
         "//tensorflow/stream_executor/gpu:redzone_allocator",
         "@com_google_absl//absl/algorithm:container",
