+++ tensorflow/core/kernels/gpu_utils.cc
@@ -24,8 +24,8 @@ limitations under the License.
 #include "tensorflow/core/platform/logger.h"
 #include "tensorflow/core/protobuf/autotuning.pb.h"
 #include "tensorflow/core/protobuf/conv_autotuning.pb.h"
+#include "tensorflow/core/util/env_var.h"
 #include "tensorflow/core/util/proto/proto_utils.h"
-#include "tensorflow/stream_executor/cuda/cuda_helpers.h"
 #include "tensorflow/stream_executor/gpu/asm_compiler.h"
 #include "tensorflow/stream_executor/gpu/redzone_allocator.h"
 
@@ -212,6 +212,22 @@ void LogFusedConvForwardAutotuneResults(
   Logger::GetSingleton()->LogProto(log);
 }
 
+// A helper function to decide whether to enable deterministic functionality.
+bool RequireDeterminism() {
+  static bool require_determinism = [] {
+    bool deterministic_ops = false;
+    TF_CHECK_OK(tensorflow::ReadBoolFromEnvVar("TF_DETERMINISTIC_OPS",
+                                               /*default_val=*/false,
+                                               &deterministic_ops));
+    bool cudnn_deterministic = false;
+    TF_CHECK_OK(tensorflow::ReadBoolFromEnvVar("TF_CUDNN_DETERMINISTIC",
+                                               /*default_val=*/false,
+                                               &cudnn_deterministic));
+    return deterministic_ops || cudnn_deterministic;
+  }();
+  return require_determinism;
+}
+
 Status BestCudnnConvAlgorithm(absl::Span<const AutotuneResult> results,
                               se::dnn::AlgorithmConfig* algo) {
   std::vector<AutotuneResult> filtered_results;
@@ -228,7 +244,7 @@ Status BestCudnnConvAlgorithm(absl::Span<const AutotuneResult> results,
 
   auto selected_result = filtered_results.begin();
   auto selected_result_no_scratch = filtered_results_no_scratch.begin();
-  if (!se::cuda::RequireCuDNNDeterminism()) {
+  if (!RequireDeterminism()) {
     auto compare_run_times = [](const AutotuneResult& lhs,
                                 const AutotuneResult& rhs) {
       return proto_utils::FromDurationProto(lhs.run_time()) <
